refactor: Delete old functions, rename and move methods
